% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibonacci.R
\name{fibo}
\alias{fibo}
\title{Compute fibonacci sequence}
\usage{
fibo(n, x = c(0, 1))
}
\description{
This function computes fibonacci sequence.
careful) Negative numbers cannot be used in n.
}
\examples{
fibo(5), fibo(125)

fibonacci_sequence()
}
