% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibonacci.R
\name{fibonacci}
\alias{fibonacci}
\title{A function to generate a Fibonacci sequences.}
\usage{
fibonacci(len)
}
\arguments{
\item{len}{numeric. Specify the length of the sequence.}
}
\description{
This function generates a Fibonacci sequence of a specified length.
}
