% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flower_match.R
\name{flower_match}
\alias{flower_match}
\title{A function to find flowers matching a Fibonacci sequence}
\usage{
flower_match(len)
}
\arguments{
\item{len}{numeric. Specify the length of the sequence.}
}
\description{
A function to find flowers matching a Fibonacci sequence
}
