\name{getFibonacci}

\alias{getFibonacci}


\title{Get Fibonacci Number}


\description{
This function returns the specified term of Fibonacci sequence.
}

\usage{
getFibonacci(n)
}

\arguments{
  \item{n}{positive integer or 0}
}  

\details{
  \code{getFibonacci} returns n'th term of Fibonacci sequence if input is correct. If n is negative returns NULL.
}

\examples{
getFibonacci(11)
}

\author{Anya Yaveyn}
