% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fibonacci.R
\name{plot_fibonacci}
\alias{plot_fibonacci}
\title{A function to plot a Fibonacci sequences.}
\usage{
plot_fibonacci(len, gradient = T)
}
\arguments{
\item{len}{numeric. Specify the length of the sequence.}

\item{gradient}{logical. Default TRUE. Indicates whether points on the plot
should be coloured by gradient.}
}
\description{
This function creates a plot for a Fibonacci sequence of a specified length.
}
