\name{printFibonacci}

\alias{printFibonacci}


\title{Print Fibonacci Number}


\description{
This function prints the specified term of Fibonacci sequence.
}

\usage{
printFibonacci(n)
}

\arguments{
  \item{n}{positive integer or 0}
}  

\details{
  \code{printFibonacci} prints n'th term of Fibonacci sequence if input is correct. If n is negative it prints error message.
}

\examples{
printFibonacci(11)
}

\author{Anya Yaveyn}
