% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fibonacci.R
\name{plot_fibonacci}
\alias{plot_fibonacci}
\title{A function to plot a Fibonacci sequences.}
\usage{
plot_fibonacci(len)
}
\arguments{
\item{len}{numeric. Specify the length of the sequence.}
}
\description{
This function creates a plot for a Fibonacci sequence of a specified length.
}
